backend/requirements.txt
# FastAPI 및 관련 패키지
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
email-validator==2.0.0

# 데이터베이스
sqlalchemy==2.0.12
alembic==1.10.4

# 보안 관련
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.6.0
bcrypt==4.0.1

# Google Cloud API
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.19.0
google-cloud-translate==3.11.1

# 텍스트 및 음성 처리
nltk==3.8.1
jellyfish==0.9.0
soundfile==0.12.1
librosa==0.10.0

# 테스트
pytest==7.3.1
httpx==0.24.0

# 기타
python-dotenv==1.0.0